+++ src/components/DownloadButton/index.tsx
@@ -243,7 +243,7 @@ export const DownloadButton: React.FC<Props> = (props) => {
       {showDropdown && (
         <div className='flex justify-center' ref={dropdownRef}>
           <div className='absolute w-full sm:w-1/2 lg:w-1/4 2xl:w-1/5 h-auto mt-2 z-10 px-6 sm:px-0'>
-            <div className='bg-[#2e2e2e] text-white border border-white/[.2] rounded-xl shadow-xl relative'>
+            <div className='bg-[#2e2e2e] overflow-hidden text-white border border-white/[.2] rounded-xl shadow-xl relative'>
               {loading ? (
                 <>
                   <div className='flex p-6 justify-center items-center'>